+++ pkg/storage/bucket/azure/bucket_client.go
@@ -4,19 +4,25 @@ import (
 	"github.com/go-kit/log"
 	"github.com/prometheus/common/model"
 	"github.com/thanos-io/objstore"
+	"github.com/thanos-io/objstore/exthttp"
 	"github.com/thanos-io/objstore/providers/azure"
-	yaml "gopkg.in/yaml.v2"
 )
 
 func NewBucketClient(cfg Config, name string, logger log.Logger) (objstore.Bucket, error) {
+	// retain default endpoint if it is not explicitly configured
+	endpoint := azure.DefaultConfig.Endpoint
+	if cfg.EndpointSuffix != "" {
+		endpoint = cfg.EndpointSuffix
+	}
+
 	bucketConfig := azure.Config{
 		StorageAccountName:      cfg.StorageAccountName,
 		StorageAccountKey:       cfg.StorageAccountKey.String(),
 		StorageConnectionString: cfg.ConnectionString.String(),
 		ContainerName:           cfg.ContainerName,
-		Endpoint:                cfg.EndpointSuffix,
+		Endpoint:                endpoint,
 		MaxRetries:              cfg.MaxRetries,
-		HTTPConfig: azure.HTTPConfig{
+		HTTPConfig: exthttp.HTTPConfig{
 			IdleConnTimeout:       model.Duration(cfg.IdleConnTimeout),
 			ResponseHeaderTimeout: model.Duration(cfg.ResponseHeaderTimeout),
 			InsecureSkipVerify:    cfg.InsecureSkipVerify,
@@ -28,12 +34,5 @@ func NewBucketClient(cfg Config, name string, logger log.Logger) (objstore.Bucke
 		},
 	}
 
-	// Thanos currently doesn't support passing the config as is, but expects a YAML,
-	// so we're going to serialize it.
-	serialized, err := yaml.Marshal(bucketConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	return azure.NewBucket(logger, serialized, name)
+	return azure.NewBucketWithConfig(logger, bucketConfig, name)
 }