+++ pkg/storage/bucket/client.go
@@ -135,6 +135,7 @@ func bucketWithMetrics(bucketClient objstore.Bucket, name string, reg prometheus
 		return bucketClient
 	}
 
+	reg = prometheus.WrapRegistererWithPrefix("loki_", reg)
 	return objstore.WrapWithMetrics(
 		bucketClient,
 		prometheus.WrapRegistererWith(prometheus.Labels{"component": name}, reg),