+++ pkg/storage/bucket/gcs/bucket_client.go
@@ -6,7 +6,6 @@ import (
 	"github.com/go-kit/log"
 	"github.com/thanos-io/objstore"
 	"github.com/thanos-io/objstore/providers/gcs"
-	yaml "gopkg.in/yaml.v2"
 )
 
 // NewBucketClient creates a new GCS bucket client
@@ -16,12 +15,5 @@ func NewBucketClient(ctx context.Context, cfg Config, name string, logger log.Lo
 		ServiceAccount: cfg.ServiceAccount.String(),
 	}
 
-	// Thanos currently doesn't support passing the config as is, but expects a YAML,
-	// so we're going to serialize it.
-	serialized, err := yaml.Marshal(bucketConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	return gcs.NewBucket(ctx, logger, serialized, name)
+	return gcs.NewBucketWithConfig(ctx, logger, bucketConfig, name)
 }